More Windows stuff

+++ numbers_windows.cpp
@@ -1,12 +1,10 @@
-// TODO: check that this actually works
-
 typedef signed char s8;
-typedef short s16;
-typedef int s32;
-typedef long long s64;
+typedef signed short s16;
+typedef signed int s32;
+typedef signed long s64;
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned u32;
-typedef unsigned long long u64;
-// defined in c:\program files (x86)\windows kits\10\include\10.0.17763.0\shared\rpcndr.h(191) when using cl
+typedef unsigned int u32;
+typedef unsigned long u64;
+// byte is defined in c:\program files (x86)\windows kits\10\include\10.0.17763.0\shared\rpcndr.h(191) when using cl
 // typedef char byte;
